requirements.txt
fastapi
uvicorn
composio-core
openai